src/backend/requirements.txt
fastapi_offline
uvicorn==0.17.6
broadcaster
asyncio_redis
logbook
websockets
fastapi-login
werkzeug
environs
cachetools
aiohttp
sentry_sdk
